+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added `overrideLock` parameter to update and delete form functions
 - docs for all available search parameters when searching for forms
 - `gatewayId` to `FormSubmissionEventConfiguration`
+- `generatePageElement` to `Forms` class
 
 # 0.7.0 (2020-06-10)